Move batch to own class.
