--- airbyte-integrations/bases/base-java/src/main/java/io/airbyte/integrations/destination_async/Batch.java
@@ -1,2 +1,0 @@
-package io.airbyte.integrations.destination_async;public class Batch {
-}
