+++ airbyte-integrations/bases/base-java/src/test/java/io/airbyte/integrations/destination_async/BufferManagerDequeueTest.java
@@ -8,8 +8,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import io.airbyte.commons.json.Jsons;
 import io.airbyte.integrations.destination_async.BufferManager.BufferManagerDequeue;
-import io.airbyte.integrations.destination_async.BufferManager.BufferManagerDequeue.Batch;
 import io.airbyte.integrations.destination_async.BufferManager.BufferManagerEnqueue;
+import io.airbyte.integrations.destination_async.buffers.Batch;
 import io.airbyte.protocol.models.v0.AirbyteMessage;
 import io.airbyte.protocol.models.v0.AirbyteMessage.Type;
 import io.airbyte.protocol.models.v0.AirbyteRecordMessage;
